Move `solve_cubic` to Fuller module and rename to `solve_fuller_cubic` since it's no longer a general purpose solver.
