+++ src/proj/fuller.rs
@@ -1,0 +1,302 @@
+//! Contains types related to the [`Fuller`] projection.
+use std::f64::consts::FRAC_PI_3;
+use super::*;
+
+/// The [Fuller](https://en.wikipedia.org/wiki/Dymaxion_map) projection method.
+///
+/// This projection preserves distances exactly along the boundaries of base triangles.
+#[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash)]
+pub struct Fuller;
+
+impl Fuller {
+    /// Constructs a [`Projection`] for a specified spherical triangle. The triangle
+    /// must be equilateral, with the angle between any two endpoints equal to `angle`.
+    pub fn triangle(self, angle: f64, points: [[f64; 3]; 3]) -> Triangle {
+        // TODO: Validate triangle when debug assertions are on
+        Triangle { angle, points }
+    }
+}
+
+impl BaseTriProjector for Fuller {
+    type Triangle = Triangle;
+    fn inside(&self, edge: basetri::HalfEdge) -> Triangle {
+        self.triangle(
+            edge.sphere().edge_length(),
+            [
+                edge.start().pos(),
+                edge.next().start().pos(),
+                edge.prev().start().pos(),
+            ],
+        )
+    }
+}
+
+/// A [`Projection`] for an equilateral triangle using the [`Fuller`] projection.
+#[derive(Clone, Copy, Debug)]
+pub struct Triangle {
+    /// The angle, in radians, between any two endpoints of the triangle.
+    angle: f64,
+    
+    /// The positions of the endpoints of the triangle on the sphere.
+    points: [[f64; 3]; 3],
+}
+
+impl Projection for Triangle {
+    fn to_sphere(&self, [u, v]: [f64; 2]) -> [f64; 3] {
+        // Special handling for boundaries to improve accuracy and performance
+        if v <= 0.0 {
+            debug_assert_eq!(v, 0.0);
+            return if u <= 0.0 {
+                debug_assert_eq!(u, 0.0);
+                self.points[0]
+            } else {
+                vec::normalize(vec::add(
+                    vec::mul(self.points[0], ((1.0 - u) * self.angle).sin()),
+                    vec::mul(self.points[1], (u * self.angle).sin()),
+                ))
+            }
+        } else if u <= 0.0 {
+            debug_assert_eq!(u, 0.0);
+            return vec::normalize(vec::add(
+                vec::mul(self.points[0], ((1.0 - v) * self.angle).sin()),
+                vec::mul(self.points[2], (v * self.angle).sin()),
+            ));
+        }
+        
+        // Adapted from "Exact Equations for Fuller’s Map Projection and Inverse"
+        // https://utppublishing.com/doi/pdf/10.3138/carto.43.1.67
+        // "Inverse with Spherical Linear Interpolation"
+        
+        // First, solve the equation:
+        // `tan(x + proj_u) + tan(x) + tan(x + proj_v) + t_alpha = 0`
+        let t_alpha = (self.angle / 2.0).tan();
+        let proj_u = (2.0 * u + v - 1.0) * self.angle;
+        let proj_v = (2.0 * v + u - 1.0) * self.angle;
+        let t_half_proj_u = (proj_u / 2.0).tan();
+        let t_half_proj_v = (proj_v / 2.0).tan();
+        let den_u = 1.0 - t_half_proj_u * t_half_proj_u;
+        let den_v = 1.0 - t_half_proj_v * t_half_proj_v;
+        let s = t_half_proj_u * den_v + t_half_proj_v * den_u;
+        let p = t_half_proj_u * t_half_proj_v;
+        let (t_x, _) = solve_cubic_fuller(
+            4.0 * p,
+            4.0 * t_alpha * p - 4.0 * s,
+            3.0 * den_u * den_v - 8.0 * p - 2.0 * t_alpha * s,
+            t_alpha * den_u * den_v + 2.0 * s,
+        )
+            .into_iter()
+            .map(|t_x| {
+                // There may be up to three solutions to the cubic equation, but only one
+                // corresponds to the correct projection solution. This is the unique solution
+                // which satisfies:
+                //   * `-self.full_angle / 2.0 <= x <= self.full_angle / 2.0`
+                //   * `-self.full_angle / 2.0 <= x + proj_u <= self.full_angle / 2.0`
+                //   * `-self.full_angle / 2.0 <= x + proj_v <= self.full_angle / 2.0`
+                //
+                // or equivalently:
+                //   * `-t_alpha <= tan(x) <= t_alpha`
+                //   * `-t_alpha <= tan(x + proj_u) <= t_alpha`
+                //   * `-t_alpha <= tan(x + proj_v) <= t_alpha`
+                
+                // Assuming there is exactly one such solution, it will be the one with the
+                // smallest `max(|tan(x)|, |tan(x + proj_u)|, |tan(x + proj_v)|)`
+                let t_x_proj_u =
+                    (den_u + 2.0 * t_half_proj_u) / (den_u - 2.0 * t_x * t_half_proj_u);
+                let t_x_proj_v =
+                    (den_v + 2.0 * t_half_proj_v) / (den_v - 2.0 * t_x * t_half_proj_v);
+                (t_x, t_x.abs().max(t_x_proj_u.abs()).max(t_x_proj_v.abs()))
+            })
+            .min_by(|(_, a), (_, b)| a.abs().partial_cmp(&b.abs()).unwrap())
+            .unwrap();
+        
+        let x = t_x.atan();
+        
+        // Use the solution to compute angle offsets of two planes from the second and
+        // third edges of the triangle
+        let a_0 = self.angle / 2.0 - x;
+        let a_1 = a_0 - (2.0 * u + v - 1.0) * self.angle;
+        
+        // Construct planes by offsetting the second and third edges of the triangle
+        let n_0 = vec::cross(
+            mat::apply(
+                [self.points[0], self.points[1]],
+                [(self.angle - a_0).sin(), a_0.sin()],
+            ),
+            vec::sub(self.points[2], self.points[1]),
+        );
+        let n_1 = vec::cross(
+            mat::apply(
+                [self.points[1], self.points[2]],
+                [(self.angle - a_1).sin(), a_1.sin()],
+            ),
+            vec::sub(self.points[0], self.points[2]),
+        );
+        
+        // The projected point is the intersection of the two planes
+        vec::normalize(vec::cross(n_0, n_1))
+    }
+    
+    fn from_sphere(&self, point: [f64; 3]) -> [f64; 2] {
+        // Adapted from "Exact Equations for Fuller’s Map Projection and Inverse"
+        // https://utppublishing.com/doi/pdf/10.3138/carto.43.1.67
+        // "Transformation Equations with Spherical Linear Interpolation"
+        
+        let d_0 = mat::det_3([point, self.points[0], self.points[1]]);
+        let d_1 = mat::det_3([point, self.points[1], self.points[2]]);
+        let d_2 = mat::det_3([point, self.points[2], self.points[0]]);
+        let c = vec::dot(self.points[0], self.points[1]);
+        let a_0 = (2.0 * (d_0 * c) / (d_0 * c + d_1 + d_2)).atan();
+        let a_1 = (2.0 * (d_1 * c) / (d_1 * c + d_2 + d_0)).atan();
+        let a_2 = (2.0 * (d_2 * c) / (d_2 * c + d_0 + d_1)).atan();
+        let u = (self.angle + 2.0 * a_2 - a_0 - a_1) / (3.0 * self.angle);
+        let v = (self.angle + 2.0 * a_0 - a_1 - a_2) / (3.0 * self.angle);
+        [u, v]
+    }
+    
+    type Transform = Transform<Self>;
+    fn transform(self, offset: [f64; 2], linear: [[f64; 2]; 2]) -> Transform<Self> {
+        Transform::new(self, offset, linear)
+    }
+}
+
+#[test]
+fn test_roundtrip() {
+    use crate::basetri::BaseTriSphere;
+    use crate::prelude::*;
+    const N: usize = 100;
+    for v in 0..N {
+        for u in 0..(N - v) {
+            let u = u as f64 / N as f64;
+            let v = v as f64 / N as f64;
+            let proj = Fuller.inside(BaseTriSphere::Icosa.face(0).side(0));
+            let point = proj.to_sphere([u, v]);
+            let [a_u, a_v] = proj.from_sphere(point);
+            assert!(
+                (a_u - u).abs() < 1e-12 && (a_v - v).abs() < 1e-12,
+                "failed roundtrip: got {:?}, expected {:?}",
+                [a_u, a_v],
+                [u, v]
+            );
+        }
+    }
+}
+
+const SMALL: f64 = 1.0e-6;
+
+/// Represents the distinct real roots of a cubic equation used in the Fuller projection. This library will only ever
+/// solve equations with 2 or 3 distinct real roots so we neglect handling other cases in the name of performance.
+#[derive(Copy, Clone)]
+struct Roots(f64, f64, Option<f64>);
+
+impl Roots {
+    /// Adds `v` to the value of all roots.
+    #[inline]
+    const fn add(self, v: f64) -> Self {
+        match self.2 {
+            None => Self(self.0 + v, self.1 + v, None),
+            Some(c) => Self(self.0 + v, self.1 + v, Some(c + v))
+        }
+    }
+}
+
+impl IntoIterator for Roots {
+    type Item = f64;
+    type IntoIter = std::iter::Take<std::array::IntoIter<f64, 3>>;
+    
+    fn into_iter(self) -> Self::IntoIter {
+        match self.2 {
+            None => [self.0, self.1, 0.0].into_iter().take(2),
+            Some(c) => [self.0, self.1, c].into_iter().take(3)
+        }
+    }
+}
+
+/// Determines the real roots of the variety of cubic polynomials which appear in the Fuller projection. If a root has
+/// multiplicity greater than one, it will only appear once in the output. Takes 36.9 FLOPs in this library's typical
+/// usage. This is not a general purpose solver and is optimized for the specific types of cubics encountered in the
+/// Fuller projection. As such, multiple branches that are never encountered in the Fuller projection have been removed
+/// for speed so this method *will* fail if used as a general purpose solver.
+fn solve_cubic_fuller(a: f64, b: f64, c: f64, d: f64) -> Roots {
+    if a.abs() < SMALL {
+        // Branch taken ~10% of the time in testing.
+        // Solve quadratic
+        let j = 2.0 * d;
+        let k = 2.0 * b;
+        let disc = c * c - j * k;
+        
+        // The polynomial discriminant will never be negative so we don't need to branch here.
+        // TODO: Can we find a proof?
+        debug_assert!(disc >= 0.0, "Quadratic discriminant was negative.");
+        let u = -c - disc.sqrt().copysign(c);
+        Roots(j / u, u / k, None)
+    } else {
+        // Convert to a depressed cubic
+        // https://en.wikipedia.org/wiki/Cubic_equation#Depressed_cubic
+        // https://mathworld.wolfram.com/CubicFormula.html
+        // In a general solver, a case for when `d` is close to 0 could improve performance. In testing this library,
+        // `d` was never smaller than 10^-6 so the branch would not be useful.
+        let b2 = b * b;
+        let a2 = a * a;
+        let ac = a * c;
+        
+        let p = (3.0 * ac - b2) / (3.0 * a2);
+        // If `b` is close to 0, this expression will be very close to `d/a`. In testing this library `|b|` was less
+        // than 10^-6 in less than 0.0001% of cases so the branch would cost more performance than it saved.
+        let q = (2.0 * b2 * b - 9.0 * ac * b + 27.0 * a2 * d) / (27.0 * a2 * a);
+        
+        let s = b / (3.0 * a);
+        
+        // Solve depressed cubic
+        // In general, a case to just output the negative cube root of `q` when `p` is close to 0 would improve performance.
+        // However, in testing, `|p|` was less than 10^-6 in less than 0.0001% of cases, so branching would cost more than
+        // it gained.
+        
+        // The discriminant will always be negative so we will always have three real roots and don't need to branch here.
+        // TODO: Can we find a proof?
+        debug_assert!(4.0 * p * p * p + 27.0 * q * q < 0.0, "Depressed cubic discriminant was not negative.");
+        let p_3 = -p / 3.0;
+        let k = 2.0 * p_3.sqrt();
+        // Absolute value not necessary after pulling p_3*p_3 out of sqrt because p is always negative.
+        let theta = (-q / (p_3 * k)).acos() / 3.0;
+        const PHI: f64 = 2.0 * FRAC_PI_3;
+        
+        Roots(
+            k * theta.cos() - s,
+            k * (theta + PHI).cos() - s,
+            Some(k * (theta + 2.0 * PHI).cos() - s)
+        )
+    }
+}
+
+#[test]
+fn test_solve_cubic() {
+    // Cubic cases
+    assert_similar(
+        solve_cubic_fuller(1.0, 4.0, 0.0, -5.0).into_iter().collect(),
+        [
+            -(5.0f64.sqrt() + 5.0) / 2.0,
+            (5.0f64.sqrt() - 5.0) / 2.0,
+            1.0,
+        ],
+    );
+    assert_similar(
+        solve_cubic_fuller(1.0, -2.0, 0.0, 1.0).into_iter().collect(),
+        [
+            (1.0 - 5.0f64.sqrt()) / 2.0,
+            1.0,
+            (1.0 + 5.0f64.sqrt()) / 2.0,
+        ],
+    );
+    
+    // Quadratic cases
+    assert_similar(solve_cubic_fuller(0.0, 1.0, 0.0, -1.0).into_iter().collect(), [-1.0, 1.0]);
+}
+
+#[cfg(test)]
+fn assert_similar<const N: usize>(mut a: Vec<f64>, mut b: [f64; N]) {
+    a.sort_by(|u, v| u.partial_cmp(v).unwrap());
+    b.sort_by(|u, v| u.partial_cmp(v).unwrap());
+    if a.len() != b.len() || a.iter().zip(b.iter()).any(|(x, y)| (x - y).abs() > SMALL) {
+        panic!("vectors are not similar, a = {:?}, b = {:?}", a, b);
+    }
+}
